Alexa-MQTT/lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
paho-mqtt==1.5.1